# [PERF] base: restore previous res.users display name search behavior