@robodoo 2024-11-08
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/186796.png)](https://mergebot.odoo.com/odoo/odoo/pull/186796)