@ryv-odoo 2024-11-08
LGTM. Sry for the issue.
@robodoo r+